Add files via upload
accNumSeq.txt fill will provide input to the account no in the application.
Bank.java file contains source code for the state like bank name, ifsc, etc and behavior of a bank like deposit, withdrow,etc.
BankAccount.java file contains the source code to create and update the details in the bank account. 
BankFrontOffice.java is the main method class which contains the different services available on the application. the operations of this bank is operate from this class.